Move and rename ExcelAdapter to PastableTable
Move the ExcelAdapter JTable structure used by the scale-matrix tool to its own class within the objects package. This essentially enables pasting with Ctrl-V to the PastableeTable object (copy functions already enabled).
