+++ src/main/java/scriptmanager/window_interface/Read_Analysis/ScaleMatrixWindow.java
@@ -5,25 +5,18 @@ import java.awt.Component;
 import java.awt.Container;
 import java.awt.Cursor;
 import java.awt.Font;
-import java.awt.Toolkit;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.DataFlavor;
-import java.awt.datatransfer.StringSelection;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
-import java.awt.event.KeyEvent;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.StringTokenizer;
 
 import javax.swing.ButtonGroup;
 import javax.swing.JButton;
-import javax.swing.JComponent;
 import javax.swing.JFileChooser;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
@@ -34,13 +27,13 @@ import javax.swing.JRadioButton;
 import javax.swing.JScrollPane;
 import javax.swing.JTable;
 import javax.swing.JTextField;
-import javax.swing.KeyStroke;
 import javax.swing.SpringLayout;
 import javax.swing.SwingConstants;
 import javax.swing.SwingWorker;
 import javax.swing.border.EmptyBorder;
 import javax.swing.table.DefaultTableModel;
 
+import scriptmanager.objects.PasteableTable;
 import scriptmanager.util.ExtensionFileFilter;
 import scriptmanager.util.FileSelection;
 import scriptmanager.scripts.Read_Analysis.ScaleMatrix;
@@ -178,7 +171,7 @@ public class ScaleMatrixWindow extends JFrame implements ActionListener, Propert
 		tableScale.setColumnSelectionAllowed(true);
 		tableScale.setRowSelectionAllowed(true);
 		@SuppressWarnings("unused")
-		ExcelAdapter myAd = new ExcelAdapter(tableScale);
+		PasteableTable myAd = new PasteableTable(tableScale);
 
 		JScrollPane scrollPane = new JScrollPane(tableScale);
 		sl_contentPane.putConstraint(SpringLayout.WEST, scrollPane, 10, SpringLayout.WEST, contentPane);
@@ -380,113 +373,6 @@ public class ScaleMatrixWindow extends JFrame implements ActionListener, Propert
 			} else {
 				lblUniformScalingFactor.setEnabled(false);
 				txtUniform.setEnabled(false);
-			}
-		}
-	}
-
-	/**
-	 * Adapted from:
-	 * https://www.javaworld.com/article/2077579/learn-java/java-tip-77--enable-copy-and-paste-functionality-between-swing-s-jtables-and-excel.html
-	 * ExcelAdapter enables Copy-Paste Clipboard functionality on JTables. The
-	 * clipboard data format used by the adapter is compatible with the clipboard
-	 * format used by Excel. This provides for clipboard interoperability between
-	 * enabled JTables and Excel.
-	 */
-	public class ExcelAdapter implements ActionListener {
-		private String rowstring, value;
-		private Clipboard system;
-		private StringSelection stsel;
-		private JTable jTable1;
-
-		/**
-		 * The Excel Adapter is constructed with a JTable on which it enables Copy-Paste
-		 * and acts as a Clipboard listener.
-		 */
-		public ExcelAdapter(JTable myJTable) {
-			jTable1 = myJTable;
-			KeyStroke copy = KeyStroke.getKeyStroke(KeyEvent.VK_C, ActionEvent.CTRL_MASK, false);
-			// Identifying the copy KeyStroke user can modify this to copy on some other Key
-			// combination.
-			KeyStroke paste = KeyStroke.getKeyStroke(KeyEvent.VK_V, ActionEvent.CTRL_MASK, false);
-			// Identifying the Paste KeyStroke user can modify this to copy on some other
-			// Key combination.
-			jTable1.registerKeyboardAction(this, "Copy", copy, JComponent.WHEN_FOCUSED);
-			jTable1.registerKeyboardAction(this, "Paste", paste, JComponent.WHEN_FOCUSED);
-			system = Toolkit.getDefaultToolkit().getSystemClipboard();
-		}
-
-		/**
-		 * Public Accessor methods for the Table on which this adapter acts.
-		 */
-		public JTable getJTable() {
-			return jTable1;
-		}
-
-		public void setJTable(JTable jTable1) {
-			this.jTable1 = jTable1;
-		}
-
-		/**
-		 * This method is activated on the Keystrokes we are listening to in this
-		 * implementation. Here it listens for Copy and Paste ActionCommands. Selections
-		 * comprising non-adjacent cells result in invalid selection and then copy
-		 * action cannot be performed. Paste is done by aligning the upper left corner
-		 * of the selection with the 1st element in the current selection of the JTable.
-		 */
-		public void actionPerformed(ActionEvent e) {
-			if (e.getActionCommand().compareTo("Copy") == 0) {
-				StringBuffer sbf = new StringBuffer();
-				// Check to ensure we have selected only a contiguous block of cells
-				int numcols = jTable1.getSelectedColumnCount();
-				int numrows = jTable1.getSelectedRowCount();
-				int[] rowsselected = jTable1.getSelectedRows();
-				int[] colsselected = jTable1.getSelectedColumns();
-				if (!((numrows - 1 == rowsselected[rowsselected.length - 1] - rowsselected[0]
-						&& numrows == rowsselected.length)
-						&& (numcols - 1 == colsselected[colsselected.length - 1] - colsselected[0]
-								&& numcols == colsselected.length))) {
-					JOptionPane.showMessageDialog(null, "Invalid Copy Selection", "Invalid Copy Selection",
-							JOptionPane.ERROR_MESSAGE);
-					return;
-				}
-				for (int i = 0; i < numrows; i++) {
-					for (int j = 0; j < numcols; j++) {
-						sbf.append(jTable1.getValueAt(rowsselected[i], colsselected[j]));
-						if (j < numcols - 1) {
-							sbf.append("\t");
-						}
-					}
-					sbf.append("\n");
-				}
-				stsel = new StringSelection(sbf.toString());
-				system = Toolkit.getDefaultToolkit().getSystemClipboard();
-				system.setContents(stsel, stsel);
-			}
-
-			if (e.getActionCommand().compareTo("Paste") == 0) {
-				// System.out.println("Trying to Paste");
-				int startRow = (jTable1.getSelectedRows())[0];
-				int startCol = (jTable1.getSelectedColumns())[0];
-				try {
-					String trstring = (String) (system.getContents(this).getTransferData(DataFlavor.stringFlavor));
-					// System.out.println("String is:"+trstring);
-					StringTokenizer st1 = new StringTokenizer(trstring, "\n");
-					for (int i = 0; st1.hasMoreTokens(); i++) {
-						rowstring = st1.nextToken();
-						StringTokenizer st2 = new StringTokenizer(rowstring, "\t");
-						for (int j = 0; st2.hasMoreTokens(); j++) {
-							value = (String) st2.nextToken();
-							if (startRow + i < jTable1.getRowCount() && startCol + j < jTable1.getColumnCount()
-									&& jTable1.isCellEditable(startRow + i, startCol + j)) {
-								jTable1.setValueAt(value, startRow + i, startCol + j);
-							}
-							// System.out.println("Putting "+ value+"at
-							// row="+startRow+i+"column="+startCol+j);
-						}
-					}
-				} catch (Exception ex) {
-					ex.printStackTrace();
-				}
 			}
 		}
 	}
